fix join counter

+++ src/counter_storage.h
@@ -1,2 +1,2 @@
 int counter_storage_init(void);
-uint32_t counter_storage_inc(void);
+uint16_t counter_storage_inc(void);
